Refactor multicluster test install
This change continues the work from #7403 by refactoring the
multicluster tests in order to install components programatically.

As part of this change, we now generate certificates (a CA and a shared
issuer) in code, and add a few utilities to manage different Kubernetes
contexts; a few examples are `KubectlApplyWithContext` and a function to
re-initialise the clientset with an arbitrary context (used only once,
since imo it's a pretty expensive operation).

Few bits and pieces have also been changed as I went through this, such
as applying entire files as opposed to reading manifests in memory
before piping them to kubectl.

Some other changes:
* remove logic from test runner script that set-up multicluster
* add a more rigurous check test after linking source to target cluster
* remove `target1` test in favour of keeping everything in one file
  (`target`)
* move manifest installation in `source` to top of the file to install
  namespaces where services will be mirrored (previously done in test
  runner)

Signed-off-by: Matei David <matei@buoyant.io>